[cassandra] Remove v1 API dependencies
