+++ internal/storage/v1/cassandra/spanstore/dbmodel/tag_filter_drop_all_test.go
@@ -7,9 +7,78 @@ import (
 	"testing"
 
 	"github.com/stretchr/testify/assert"
+
+	"github.com/jaegertracing/jaeger-idl/model/v1"
 )
 
 var _ TagFilter = &TagFilterDropAll{} // Check API compliance
+
+var (
+	someSpanID        = int64(3333)
+	someParentSpanID  = int64(11111)
+	someOperationName = "someOperationName"
+	someStartTime     = model.EpochMicrosecondsAsTime(55555)
+	someDuration      = model.MicrosecondsAsDuration(50000)
+	someFlags         = int32(1)
+	someLogTimestamp  = model.EpochMicrosecondsAsTime(12345)
+	someServiceName   = "someServiceName"
+
+	someStringTagValue = "someTagValue"
+	someBoolTagValue   = true
+	someLongTagValue   = int64(123)
+	someDoubleTagValue = float64(1.4)
+	someBinaryTagValue = []byte("someBinaryValue")
+	someStringTagKey   = "someStringTag"
+	someBoolTagKey     = "someBoolTag"
+	someLongTagKey     = "someLongTag"
+	someDoubleTagKey   = "someDoubleTag"
+	someBinaryTagKey   = "someBinaryTag"
+	someDBTags         = []KeyValue{
+		{
+			Key:         someStringTagKey,
+			ValueType:   StringType,
+			ValueString: someStringTagValue,
+		},
+		{
+			Key:       someBoolTagKey,
+			ValueType: BoolType,
+			ValueBool: someBoolTagValue,
+		},
+		{
+			Key:        someLongTagKey,
+			ValueType:  Int64Type,
+			ValueInt64: someLongTagValue,
+		},
+		{
+			Key:          someDoubleTagKey,
+			ValueType:    Float64Type,
+			ValueFloat64: someDoubleTagValue,
+		},
+		{
+			Key:         someBinaryTagKey,
+			ValueType:   BinaryType,
+			ValueBinary: someBinaryTagValue,
+		},
+	}
+	someDBLogs = []Log{
+		{
+			Timestamp: int64(model.TimeAsEpochMicroseconds(someLogTimestamp)),
+			Fields:    someDBTags,
+		},
+	}
+	someDBProcess = Process{
+		ServiceName: someServiceName,
+		Tags:        someDBTags,
+	}
+	someDBTraceID = TraceID{2, 2, 2, 2, 2, 4, 4, 4, 4}
+	someDBRefs    = []SpanRef{
+		{
+			RefType: "child-of",
+			SpanID:  someParentSpanID,
+			TraceID: someDBTraceID,
+		},
+	}
+)
 
 func TestDropAll(t *testing.T) {
 	tt := []struct {
