+++ internal/storage/v1/cassandra/spanstore/mocks/mocks.go
@@ -13,7 +13,6 @@ import (
 	"context"
 	"iter"
 
-	"github.com/jaegertracing/jaeger/internal/storage/v1/api/spanstore"
 	"github.com/jaegertracing/jaeger/internal/storage/v1/cassandra/spanstore/dbmodel"
 	"github.com/jaegertracing/jaeger/internal/storage/v2/api/tracestore"
 	mock "github.com/stretchr/testify/mock"
@@ -47,7 +46,7 @@ func (_m *CoreSpanReader) EXPECT() *CoreSpanReader_Expecter {
 }
 
 // FindTraceIDs provides a mock function for the type CoreSpanReader
-func (_mock *CoreSpanReader) FindTraceIDs(ctx context.Context, traceQuery *spanstore.TraceQueryParameters) ([]dbmodel.TraceID, error) {
+func (_mock *CoreSpanReader) FindTraceIDs(ctx context.Context, traceQuery *tracestore.TraceQueryParams) ([]dbmodel.TraceID, error) {
 	ret := _mock.Called(ctx, traceQuery)
 
 	if len(ret) == 0 {
@@ -56,17 +55,17 @@ func (_mock *CoreSpanReader) FindTraceIDs(ctx context.Context, traceQuery *spans
 
 	var r0 []dbmodel.TraceID
 	var r1 error
-	if returnFunc, ok := ret.Get(0).(func(context.Context, *spanstore.TraceQueryParameters) ([]dbmodel.TraceID, error)); ok {
+	if returnFunc, ok := ret.Get(0).(func(context.Context, *tracestore.TraceQueryParams) ([]dbmodel.TraceID, error)); ok {
 		return returnFunc(ctx, traceQuery)
 	}
-	if returnFunc, ok := ret.Get(0).(func(context.Context, *spanstore.TraceQueryParameters) []dbmodel.TraceID); ok {
+	if returnFunc, ok := ret.Get(0).(func(context.Context, *tracestore.TraceQueryParams) []dbmodel.TraceID); ok {
 		r0 = returnFunc(ctx, traceQuery)
 	} else {
 		if ret.Get(0) != nil {
 			r0 = ret.Get(0).([]dbmodel.TraceID)
 		}
 	}
-	if returnFunc, ok := ret.Get(1).(func(context.Context, *spanstore.TraceQueryParameters) error); ok {
+	if returnFunc, ok := ret.Get(1).(func(context.Context, *tracestore.TraceQueryParams) error); ok {
 		r1 = returnFunc(ctx, traceQuery)
 	} else {
 		r1 = ret.Error(1)
@@ -81,20 +80,20 @@ type CoreSpanReader_FindTraceIDs_Call struct {
 
 // FindTraceIDs is a helper method to define mock.On call
 //   - ctx context.Context
-//   - traceQuery *spanstore.TraceQueryParameters
+//   - traceQuery *tracestore.TraceQueryParams
 func (_e *CoreSpanReader_Expecter) FindTraceIDs(ctx interface{}, traceQuery interface{}) *CoreSpanReader_FindTraceIDs_Call {
 	return &CoreSpanReader_FindTraceIDs_Call{Call: _e.mock.On("FindTraceIDs", ctx, traceQuery)}
 }
 
-func (_c *CoreSpanReader_FindTraceIDs_Call) Run(run func(ctx context.Context, traceQuery *spanstore.TraceQueryParameters)) *CoreSpanReader_FindTraceIDs_Call {
+func (_c *CoreSpanReader_FindTraceIDs_Call) Run(run func(ctx context.Context, traceQuery *tracestore.TraceQueryParams)) *CoreSpanReader_FindTraceIDs_Call {
 	_c.Call.Run(func(args mock.Arguments) {
 		var arg0 context.Context
 		if args[0] != nil {
 			arg0 = args[0].(context.Context)
 		}
-		var arg1 *spanstore.TraceQueryParameters
+		var arg1 *tracestore.TraceQueryParams
 		if args[1] != nil {
-			arg1 = args[1].(*spanstore.TraceQueryParameters)
+			arg1 = args[1].(*tracestore.TraceQueryParams)
 		}
 		run(
 			arg0,
@@ -109,13 +108,13 @@ func (_c *CoreSpanReader_FindTraceIDs_Call) Return(traceIDs []dbmodel.TraceID, e
 	return _c
 }
 
-func (_c *CoreSpanReader_FindTraceIDs_Call) RunAndReturn(run func(ctx context.Context, traceQuery *spanstore.TraceQueryParameters) ([]dbmodel.TraceID, error)) *CoreSpanReader_FindTraceIDs_Call {
+func (_c *CoreSpanReader_FindTraceIDs_Call) RunAndReturn(run func(ctx context.Context, traceQuery *tracestore.TraceQueryParams) ([]dbmodel.TraceID, error)) *CoreSpanReader_FindTraceIDs_Call {
 	_c.Call.Return(run)
 	return _c
 }
 
 // FindTraces provides a mock function for the type CoreSpanReader
-func (_mock *CoreSpanReader) FindTraces(ctx context.Context, traceQuery *spanstore.TraceQueryParameters) iter.Seq2[dbmodel.Trace, error] {
+func (_mock *CoreSpanReader) FindTraces(ctx context.Context, traceQuery *tracestore.TraceQueryParams) iter.Seq2[dbmodel.Trace, error] {
 	ret := _mock.Called(ctx, traceQuery)
 
 	if len(ret) == 0 {
@@ -123,7 +122,7 @@ func (_mock *CoreSpanReader) FindTraces(ctx context.Context, traceQuery *spansto
 	}
 
 	var r0 iter.Seq2[dbmodel.Trace, error]
-	if returnFunc, ok := ret.Get(0).(func(context.Context, *spanstore.TraceQueryParameters) iter.Seq2[dbmodel.Trace, error]); ok {
+	if returnFunc, ok := ret.Get(0).(func(context.Context, *tracestore.TraceQueryParams) iter.Seq2[dbmodel.Trace, error]); ok {
 		r0 = returnFunc(ctx, traceQuery)
 	} else {
 		if ret.Get(0) != nil {
@@ -140,20 +139,20 @@ type CoreSpanReader_FindTraces_Call struct {
 
 // FindTraces is a helper method to define mock.On call
 //   - ctx context.Context
-//   - traceQuery *spanstore.TraceQueryParameters
+//   - traceQuery *tracestore.TraceQueryParams
 func (_e *CoreSpanReader_Expecter) FindTraces(ctx interface{}, traceQuery interface{}) *CoreSpanReader_FindTraces_Call {
 	return &CoreSpanReader_FindTraces_Call{Call: _e.mock.On("FindTraces", ctx, traceQuery)}
 }
 
-func (_c *CoreSpanReader_FindTraces_Call) Run(run func(ctx context.Context, traceQuery *spanstore.TraceQueryParameters)) *CoreSpanReader_FindTraces_Call {
+func (_c *CoreSpanReader_FindTraces_Call) Run(run func(ctx context.Context, traceQuery *tracestore.TraceQueryParams)) *CoreSpanReader_FindTraces_Call {
 	_c.Call.Run(func(args mock.Arguments) {
 		var arg0 context.Context
 		if args[0] != nil {
 			arg0 = args[0].(context.Context)
 		}
-		var arg1 *spanstore.TraceQueryParameters
+		var arg1 *tracestore.TraceQueryParams
 		if args[1] != nil {
-			arg1 = args[1].(*spanstore.TraceQueryParameters)
+			arg1 = args[1].(*tracestore.TraceQueryParams)
 		}
 		run(
 			arg0,
@@ -168,7 +167,7 @@ func (_c *CoreSpanReader_FindTraces_Call) Return(seq2 iter.Seq2[dbmodel.Trace, e
 	return _c
 }
 
-func (_c *CoreSpanReader_FindTraces_Call) RunAndReturn(run func(ctx context.Context, traceQuery *spanstore.TraceQueryParameters) iter.Seq2[dbmodel.Trace, error]) *CoreSpanReader_FindTraces_Call {
+func (_c *CoreSpanReader_FindTraces_Call) RunAndReturn(run func(ctx context.Context, traceQuery *tracestore.TraceQueryParams) iter.Seq2[dbmodel.Trace, error]) *CoreSpanReader_FindTraces_Call {
 	_c.Call.Return(run)
 	return _c
 }
