+++ internal/storage/v1/cassandra/spanstore/reader.go
@@ -20,7 +20,6 @@ import (
 	"github.com/jaegertracing/jaeger/internal/metrics"
 	"github.com/jaegertracing/jaeger/internal/storage/cassandra"
 	casmetrics "github.com/jaegertracing/jaeger/internal/storage/cassandra/metrics"
-	"github.com/jaegertracing/jaeger/internal/storage/v1/api/spanstore"
 	"github.com/jaegertracing/jaeger/internal/storage/v1/cassandra/spanstore/dbmodel"
 	"github.com/jaegertracing/jaeger/internal/storage/v2/api/tracestore"
 	"github.com/jaegertracing/jaeger/internal/telemetry/otelsemconv"
@@ -105,8 +104,8 @@ type CoreSpanReader interface {
 	GetServices(ctx context.Context) ([]string, error)
 	GetOperations(ctx context.Context, query tracestore.OperationQueryParams) ([]tracestore.Operation, error)
 	GetTrace(ctx context.Context, traceID dbmodel.TraceID) ([]dbmodel.Span, error)
-	FindTraces(ctx context.Context, traceQuery *spanstore.TraceQueryParameters) iter.Seq2[dbmodel.Trace, error]
-	FindTraceIDs(ctx context.Context, traceQuery *spanstore.TraceQueryParameters) ([]dbmodel.TraceID, error)
+	FindTraces(ctx context.Context, traceQuery *tracestore.TraceQueryParams) iter.Seq2[dbmodel.Trace, error]
+	FindTraceIDs(ctx context.Context, traceQuery *tracestore.TraceQueryParams) ([]dbmodel.TraceID, error)
 }
 
 // SpanReader can query for and load traces from Cassandra.
@@ -215,11 +214,11 @@ func (s *SpanReader) GetTrace(ctx context.Context, traceID dbmodel.TraceID) ([]d
 	return s.readTrace(ctx, traceID)
 }
 
-func validateQuery(p *spanstore.TraceQueryParameters) error {
+func validateQuery(p *tracestore.TraceQueryParams) error {
 	if p == nil {
 		return ErrMalformedRequestObject
 	}
-	if p.ServiceName == "" && len(p.Tags) > 0 {
+	if p.ServiceName == "" && p.Attributes.Len() > 0 {
 		return ErrServiceNameNotSet
 	}
 	if p.StartTimeMin.IsZero() || p.StartTimeMax.IsZero() {
@@ -231,14 +230,14 @@ func validateQuery(p *spanstore.TraceQueryParameters) error {
 	if p.DurationMin != 0 && p.DurationMax != 0 && p.DurationMin > p.DurationMax {
 		return ErrDurationMinGreaterThanMax
 	}
-	if (p.DurationMin != 0 || p.DurationMax != 0) && len(p.Tags) > 0 {
+	if (p.DurationMin != 0 || p.DurationMax != 0) && p.Attributes.Len() > 0 {
 		return ErrDurationAndTagQueryNotSupported
 	}
 	return nil
 }
 
 // FindTraces retrieves traces that match the traceQuery
-func (s *SpanReader) FindTraces(ctx context.Context, traceQuery *spanstore.TraceQueryParameters) iter.Seq2[dbmodel.Trace, error] {
+func (s *SpanReader) FindTraces(ctx context.Context, traceQuery *tracestore.TraceQueryParams) iter.Seq2[dbmodel.Trace, error] {
 	return func(yield func(dbmodel.Trace, error) bool) {
 		dbTraceIDs, err := s.FindTraceIDs(ctx, traceQuery)
 		if err != nil {
@@ -262,12 +261,12 @@ func (s *SpanReader) FindTraces(ctx context.Context, traceQuery *spanstore.Trace
 }
 
 // FindTraceIDs retrieve traceIDs that match the traceQuery
-func (s *SpanReader) FindTraceIDs(ctx context.Context, traceQuery *spanstore.TraceQueryParameters) ([]dbmodel.TraceID, error) {
+func (s *SpanReader) FindTraceIDs(ctx context.Context, traceQuery *tracestore.TraceQueryParams) ([]dbmodel.TraceID, error) {
 	if err := validateQuery(traceQuery); err != nil {
 		return nil, err
 	}
-	if traceQuery.NumTraces == 0 {
-		traceQuery.NumTraces = defaultNumTraces
+	if traceQuery.SearchDepth == 0 {
+		traceQuery.SearchDepth = defaultNumTraces
 	}
 
 	dbTraceIDs, err := s.findTraceIDsFromQuery(ctx, traceQuery)
@@ -277,7 +276,7 @@ func (s *SpanReader) FindTraceIDs(ctx context.Context, traceQuery *spanstore.Tra
 
 	var traceIDs []dbmodel.TraceID
 	for t := range dbTraceIDs {
-		if len(traceIDs) >= traceQuery.NumTraces {
+		if len(traceIDs) >= traceQuery.SearchDepth {
 			break
 		}
 		traceIDs = append(traceIDs, t)
@@ -285,7 +284,7 @@ func (s *SpanReader) FindTraceIDs(ctx context.Context, traceQuery *spanstore.Tra
 	return traceIDs, nil
 }
 
-func (s *SpanReader) findTraceIDsFromQuery(ctx context.Context, traceQuery *spanstore.TraceQueryParameters) (dbmodel.UniqueTraceIDs, error) {
+func (s *SpanReader) findTraceIDsFromQuery(ctx context.Context, traceQuery *tracestore.TraceQueryParams) (dbmodel.UniqueTraceIDs, error) {
 	// See docs/adr/cassandra-find-traces-duration.md for rationale: duration queries use the duration_index
 	// and are handled as a separate path. Other query parameters (like tags) are ignored when duration is specified.
 	if traceQuery.DurationMin != 0 || traceQuery.DurationMax != 0 {
@@ -297,7 +296,7 @@ func (s *SpanReader) findTraceIDsFromQuery(ctx context.Context, traceQuery *span
 		if err != nil {
 			return nil, err
 		}
-		if len(traceQuery.Tags) > 0 {
+		if traceQuery.Attributes.Len() > 0 {
 			tagTraceIds, err := s.queryByTagsAndLogs(ctx, traceQuery)
 			if err != nil {
 				return nil, err
@@ -309,22 +308,22 @@ func (s *SpanReader) findTraceIDsFromQuery(ctx context.Context, traceQuery *span
 		}
 		return traceIds, nil
 	}
-	if len(traceQuery.Tags) > 0 {
+	if traceQuery.Attributes.Len() > 0 {
 		return s.queryByTagsAndLogs(ctx, traceQuery)
 	}
 	return s.queryByService(ctx, traceQuery)
 }
 
-func (s *SpanReader) queryByTagsAndLogs(ctx context.Context, tq *spanstore.TraceQueryParameters) (dbmodel.UniqueTraceIDs, error) {
+func (s *SpanReader) queryByTagsAndLogs(ctx context.Context, tq *tracestore.TraceQueryParams) (dbmodel.UniqueTraceIDs, error) {
 	ctx, span := s.startSpanForQuery(ctx, "queryByTagsAndLogs", queryByTag)
 	defer span.End()
 
-	results := make([]dbmodel.UniqueTraceIDs, 0, len(tq.Tags))
-	for k, v := range tq.Tags {
+	results := make([]dbmodel.UniqueTraceIDs, 0, tq.Attributes.Len())
+	for k, v := range tq.Attributes.All() {
 		_, childSpan := s.tracer.Start(ctx, "queryByTag")
 		childSpan.SetAttributes(
 			attribute.Key("tag.key").String(k),
-			attribute.Key("tag.value").String(v),
+			attribute.Key("tag.value").String(v.AsString()),
 		)
 		query := s.session.Query(
 			queryByTag,
@@ -333,7 +332,7 @@ func (s *SpanReader) queryByTagsAndLogs(ctx context.Context, tq *spanstore.Trace
 			v,
 			model.TimeAsEpochMicroseconds(tq.StartTimeMin),
 			model.TimeAsEpochMicroseconds(tq.StartTimeMax),
-			tq.NumTraces*limitMultiple,
+			tq.SearchDepth*limitMultiple,
 		).PageSize(0)
 		t, err := s.executeQuery(childSpan, query, s.metrics.queryTagIndex)
 		childSpan.End()
@@ -345,7 +344,7 @@ func (s *SpanReader) queryByTagsAndLogs(ctx context.Context, tq *spanstore.Trace
 	return dbmodel.IntersectTraceIDs(results), nil
 }
 
-func (s *SpanReader) queryByDuration(ctx context.Context, traceQuery *spanstore.TraceQueryParameters) (dbmodel.UniqueTraceIDs, error) {
+func (s *SpanReader) queryByDuration(ctx context.Context, traceQuery *tracestore.TraceQueryParams) (dbmodel.UniqueTraceIDs, error) {
 	ctx, span := s.startSpanForQuery(ctx, "queryByDuration", queryByDuration)
 	defer span.End()
 
@@ -372,7 +371,7 @@ func (s *SpanReader) queryByDuration(ctx context.Context, traceQuery *spanstore.
 			traceQuery.OperationName,
 			minDurationMicros,
 			maxDurationMicros,
-			traceQuery.NumTraces*limitMultiple)
+			traceQuery.SearchDepth*limitMultiple)
 		t, err := s.executeQuery(childSpan, query, s.metrics.queryDurationIndex)
 		childSpan.End()
 		if err != nil {
@@ -381,7 +380,7 @@ func (s *SpanReader) queryByDuration(ctx context.Context, traceQuery *spanstore.
 
 		for traceID := range t {
 			results.Add(traceID)
-			if len(results) == traceQuery.NumTraces {
+			if len(results) == traceQuery.SearchDepth {
 				break
 			}
 		}
@@ -389,7 +388,7 @@ func (s *SpanReader) queryByDuration(ctx context.Context, traceQuery *spanstore.
 	return results, nil
 }
 
-func (s *SpanReader) queryByServiceNameAndOperation(ctx context.Context, tq *spanstore.TraceQueryParameters) (dbmodel.UniqueTraceIDs, error) {
+func (s *SpanReader) queryByServiceNameAndOperation(ctx context.Context, tq *tracestore.TraceQueryParams) (dbmodel.UniqueTraceIDs, error) {
 	_, span := s.startSpanForQuery(ctx, "queryByServiceNameAndOperation", queryByServiceAndOperationName)
 	defer span.End()
 	query := s.session.Query(
@@ -398,12 +397,12 @@ func (s *SpanReader) queryByServiceNameAndOperation(ctx context.Context, tq *spa
 		tq.OperationName,
 		model.TimeAsEpochMicroseconds(tq.StartTimeMin),
 		model.TimeAsEpochMicroseconds(tq.StartTimeMax),
-		tq.NumTraces*limitMultiple,
+		tq.SearchDepth*limitMultiple,
 	).PageSize(0)
 	return s.executeQuery(span, query, s.metrics.queryServiceOperationIndex)
 }
 
-func (s *SpanReader) queryByService(ctx context.Context, tq *spanstore.TraceQueryParameters) (dbmodel.UniqueTraceIDs, error) {
+func (s *SpanReader) queryByService(ctx context.Context, tq *tracestore.TraceQueryParams) (dbmodel.UniqueTraceIDs, error) {
 	_, span := s.startSpanForQuery(ctx, "queryByService", queryByServiceAndOperationName)
 	defer span.End()
 	query := s.session.Query(
@@ -411,7 +410,7 @@ func (s *SpanReader) queryByService(ctx context.Context, tq *spanstore.TraceQuer
 		tq.ServiceName,
 		model.TimeAsEpochMicroseconds(tq.StartTimeMin),
 		model.TimeAsEpochMicroseconds(tq.StartTimeMax),
-		tq.NumTraces*limitMultiple,
+		tq.SearchDepth*limitMultiple,
 	).PageSize(0)
 	return s.executeQuery(span, query, s.metrics.queryServiceNameIndex)
 }
